debug

+++ 02_loops/08_reverse_number.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std; 
 int main() {
-    int n;
+    long long n;                   // integer value may be big, depends on user input, so better to use long long
     cout << "enter a number : ";
     cin >> n;
     if (n == 0) cout << n;     // edge case 0
